backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
numba==0.58.1
numpy==1.26.3
pydantic==2.6.1
alpaca-py
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.27
python-dotenv==1.0.0
oauthlib==3.2.2
requests-oauthlib==1.3.1 
httpx 
email-validator==2.1.0.post1
pandas==2.2.0
python-dateutil==2.8.2 